Handle GetFeeStats Data Response For Usage
